pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Sato Shingo <shinngoll@hotmail.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
moviepy = "^1.0.3"
openai = "^1.35.3"
pillow = "^10.3.0"
opencv-python = "^4.10.0.84"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
